+++ CHANGELOG.md
@@ -9,8 +9,8 @@ script to proceed.
 - Removed parameter of `temboard --debug` CLI option.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
 - Move `temboard-agent-register` as `temboard-agent register` subcommand.
-- New subcommands `temboard runtask` and `temboard-agent runtask` for
-  debugging.
+- New subcommands `temboard runtask`, `temboard-agent routes` and
+  `temboard-agent runtask` for debugging.
 
 
 ## [7.11] - Unreleased